Adding code showing how to update insights module from action plugin
The thinking here is that we replace the existing insights.py on the
local disk with what we download from Red Hat.  The big disadvantage is
that everything has to be in one big file.  This is a limitation of
Ansible:

> Modules must be self-contained in one file to be auto-transferred by ansible

(http://docs.ansible.com/ansible/dev_guide/developing_modules_best_practices.html)

Perhaps we can do a manual transfer, but I'm not sure of the
compatibility implications, and it kind of reeks of a hack.